Refactoring import statements. Now there is coherence between jupyter notebook and the package modules. Minor changes in settings. Environment updated (python version upgrade).
