+++ src/tests/test_file_manager.py
@@ -1,7 +1,8 @@
 import os
 import tempfile
 import pytest
-from util.file_manager import FileManager
+
+from src.util.file_manager import FileManager
 
 
 def test_load_file():
